+++ photutils/segmentation/properties.py
@@ -23,7 +23,6 @@ from ..aperture import (BoundingBox, CircularAperture, EllipticalAperture,
                         RectangularAnnulus)
 from ..utils._convolution import _filter_data
 from ..utils._moments import _moments, _moments_central
-from ..utils._wcs_helpers import _pixel_to_world
 
 __all__ = ['SourceProperties', 'source_properties', 'LegacySourceCatalog']
 
@@ -706,8 +705,10 @@ class SourceProperties:
         The output coordinate frame is the same as the input WCS.
         """
 
-        return _pixel_to_world(self.xcentroid.value, self.ycentroid.value,
-                               self._wcs)
+        if self._wcs is None:
+            return None
+        return self._wcs.pixel_to_world(self.xcentroid.value,
+                                        self.ycentroid.value)
 
     @lazyproperty
     def sky_centroid_icrs(self):
@@ -1997,8 +1998,8 @@ class LegacySourceCatalog:
             # SkyCoord array from a loop-generated SkyCoord list.  The
             # assumption here is that the wcs is the same for each
             # SourceProperties instance.
-            return _pixel_to_world(self.xcentroid.value, self.ycentroid.value,
-                                   self.wcs)
+            return self.wcs.pixel_to_world(self.xcentroid.value,
+                                           self.ycentroid.value)
 
     @lazyproperty
     def sky_centroid_icrs(self):
@@ -2169,6 +2170,8 @@ def _calc_sky_bbox_corner(bbox, corner, wcs):
         The sky coordinate at the bounding box corner.  If ``wcs`` is
         `None`, then `None` will be returned.
     """
+    if wcs is None:
+        return None
 
     if corner == 'll':
         xpos = bbox.ixmin - 0.5
@@ -2185,4 +2188,4 @@ def _calc_sky_bbox_corner(bbox, corner, wcs):
     else:
         raise ValueError('Invalid corner name.')
 
-    return _pixel_to_world(xpos, ypos, wcs)
+    return wcs.pixel_to_world(xpos, ypos)